+++ CarzyCarSpringBootServer/src/main/java/com/tastsong/crazycar/model/UserLoginRecordModel.java
@@ -2,7 +2,7 @@ package com.tastsong.crazycar.model;
 
 public class UserLoginRecordModel {
     public String user_name;
-    public Integer login_time;
+    public Long login_time;
     public String place;
     public String device;
 }
