+++ CarzyCarSpringBootServer/src/main/java/com/tastsong/crazycar/model/UserModel.java
@@ -4,7 +4,7 @@ public class UserModel {
     public Integer uid;
     public String user_name;
     public String user_password;
-    public Integer login_time;
+    public Long login_time;
     public Integer aid;
     public Integer star;
     public boolean is_vip;
